[~] fix liquibase settings
